refactor: Stepper의 RecoilRoot 의존성 제거
