+++ src/components/pages/ProductsPage/ProductStepper/ProductStepper.tsx
@@ -4,25 +4,36 @@ import * as Styled from './ProductStepper.styled';
 
 import StepperEntryButton from '../StepperEntryButton/StepperEntryButton';
 import Stepper from '../../../commons/Stepper/Stepper';
-import useMyCartUpdater from '../../../../hooks/useMyCartUpdater';
+import useStepper from '../../../../hooks/useStepper';
 
 import StepperSettings from '../../../../constants/StepperSettings';
 
 interface ProductStepperProps {
   productId: number;
+  defaultValue: number;
 }
 
 const { MIN, MAX, STEP } = StepperSettings;
 
 const ProductStepper = (props: ProductStepperProps) => {
-  const { productId } = props;
+  const { productId, defaultValue } = props;
 
-  const { value, increaseValue } = useMyCartUpdater(productId, { min: MIN, max: MAX, step: STEP });
+  const { value, increaseValue, decreaseValue, setValue } = useStepper(
+    MIN,
+    MAX,
+    STEP,
+    defaultValue
+  );
 
   return (
     <Styled.ProductStepper>
       {value ? (
-        <Stepper productId={productId} min={MIN} max={MAX} step={STEP} />
+        <Stepper
+          value={value}
+          increaseValue={increaseValue}
+          decreaseValue={decreaseValue}
+          setValue={setValue}
+        />
       ) : (
         <StepperEntryButton onClick={increaseValue} />
       )}
